chore(GuideDetail): split up components
