+++ src/components/GuideDetailView/GuideDetailControls/_guide-detail-controls.scss
@@ -1,0 +1,7 @@
+@import "../../../styles/responsive";
+
+.guide-controls {
+  &__button {
+    margin-left: 8px;
+  }
+}
