+++ src/components/GuideDetailView/GuideDetailSection/_guide-detail-section.scss
@@ -1,0 +1,71 @@
+@import "../../../styles/responsive";
+@import "@elastic/eui/src/themes/eui-amsterdam/eui_amsterdam_colors_dark.scss";
+
+.guide-section {
+  width: 100%;
+  margin-bottom: 14px;
+
+  &__title {
+    display: flex;
+    align-items: center;
+    padding: 8px 16px;
+    background-color: lighten($color: $euiColorEmptyShade, $amount: 4%);
+    font-size: 24px;
+    font-weight: 600;
+    text-transform: capitalize;
+    border-bottom: 1px solid $euiColorLightShade;
+
+    .euiFormControlLayout {
+      max-width: 200px;
+
+      @include respond-to(gt-sm) {
+        max-width: 400px;
+      }
+    }
+
+    &--collapse {
+      margin-left: auto;
+    }
+  }
+
+  &__body {
+    padding: 16px;
+
+    &.editing {
+      padding: 0;
+
+      .euiMarkdownEditorToolbar {
+        border-radius: 0;
+      }
+
+      .euiMarkdownEditorPreview,
+      .euiMarkdownEditorFooter {
+        border-radius: 0;
+      }
+
+      .euiMarkdownEditorFooter {
+        border-radius: 0;
+      }
+    }
+
+    @include respond-to(gt-xs) {
+      padding: 16px;
+    }
+
+    .euiMarkdownFormat {
+      p {
+        display: flex;
+        flex-direction: column;
+
+        @include respond-to(gt-sm) {
+          flex-direction: row;
+          flex-wrap: wrap;
+        }
+      }
+
+      img {
+        margin: 4px 2px;
+      }
+    }
+  }
+}
