+++ src/components/GuideDetailView/GuideDetailView.tsx
@@ -2,14 +2,6 @@ import React, { FunctionComponent, useState, ReactElement } from "react";
 
 /* Components */
 import {
-  EuiPanel,
-  EuiMarkdownEditor,
-  EuiMarkdownFormat,
-  EuiSideNav,
-  EuiButton,
-  EuiFieldText,
-  EuiAvatar,
-  EuiButtonIcon,
   EuiDragDropContext,
   EuiDroppable,
   EuiDraggable,
@@ -22,72 +14,20 @@ import "./_guide-detail-view.scss";
 
 /* Types */
 import { Guide } from "../../models/Guide";
-import { Tag } from "../../models/Tag";
 
 /* Constants */
-import { newSection } from "../../constants/constants";
+import { mockGuide, newSection } from "../../constants/constants";
+import { GuideDetailSideNav } from "./GuideDetailSideNav/GuideDetailSideNav";
+import { GuideDetailSection } from "./GuideDetailSection/GuideDetailSection";
+import { GuideDetailControls } from "./GuideDetailControls/GuideDetailControls";
 
 export interface GuideDetailViewProps {}
 
 export const GuideDetailView: FunctionComponent<GuideDetailViewProps> = (): ReactElement => {
-  const mockGuide: Guide = {
-    _id: "mock_id",
-    title: "falco",
-    sections: [
-      {
-        _id: "mock_post_id",
-        title: "basics",
-        body: `The first key to understanding how to fight falco is that both of his primary walling options (bair and utilt) have virtually the exact same range. Meaning, if you're spacing for one you're simultaneously spacing for the other. This makes it far simpler to smother him/punish him
-\n![](https://media.giphy.com/media/ZpLzabCMomHUQPbcvg/giphy.gif) ![](https://media.giphy.com/media/ZpLzabCMomHUQPbcvg/giphy.gif) ![](https://media.giphy.com/media/ZpLzabCMomHUQPbcvg/giphy.gif)`,
-        tags: Array<Tag>(),
-      },
-      {
-        _id: "mock_post_id",
-        title: "percents",
-        body: `## \`No DI\`\n
-* \`40%\` Uthrow regrab
-* \`72%\` Uthrow fsmash
-* \`85%\` Uthrow knee
-        
-## \`DI down and away\` 
-*(rough %s as it depends on notch position)*
-* \`~65%\` Uthrow regrab
-* \`~65%\` Uthrow dash SH uair
-* \`~105%\` Uthrow dash SH knee`,
-        tags: Array<Tag>(),
-      },
-      {
-        _id: "mock_post_id",
-        title: "defense-and-recovery",
-        body: `![](https://media.giphy.com/media/ZpLzabCMomHUQPbcvg/giphy.gif) ![](https://media.giphy.com/media/ZpLzabCMomHUQPbcvg/giphy.gif) ![](https://media.giphy.com/media/ZpLzabCMomHUQPbcvg/giphy.gif) `,
-        tags: Array<Tag>(),
-      },
-    ],
-    tags: Array<Tag>(),
-  };
   const [editing, setEditing] = useState<boolean>(false);
   const [collapsed, setCollapsed] = useState<Array<boolean>>([]);
   const [sections, setSections] = useState<any>(mockGuide.sections);
   const [guide, setGuide] = useState<Guide | null>(mockGuide);
-
-  const buildSideNaveItems = () => {
-    if (!guide) return;
-    return sections.map((section, index) => {
-      const { title } = section;
-      return {
-        name: title,
-        id: index,
-        onClick: () => {
-          let div = document.getElementById(title);
-          if (div && document) {
-            let topPos = div.offsetTop - 200;
-            let sectionsDiv = document.getElementById("sections");
-            if (sectionsDiv) sectionsDiv.scrollTop = topPos;
-          }
-        },
-      };
-    });
-  };
 
   const updateSection = (key, value, index) => {
     if (!guide) return;
@@ -96,7 +36,11 @@ export const GuideDetailView: FunctionComponent<GuideDetailViewProps> = (): Reac
     setGuide({ ...guide });
   };
 
-  const addSection = () => {
+  const handleSetEditing = (isEditing) => {
+    setEditing(isEditing);
+  };
+
+  const handleAddSection = () => {
     if (!guide) return;
     const { sections } = guide;
     sections.unshift(newSection);
@@ -145,52 +89,6 @@ export const GuideDetailView: FunctionComponent<GuideDetailViewProps> = (): Reac
       const { title, body } = section;
       const isCollapsed = collapsed[index] && collapsed[index] === true;
 
-      const sectionPanel = (
-        <EuiPanel
-          id={section.title}
-          hasShadow={false}
-          hasBorder
-          className="guide-section"
-          key={index}
-        >
-          <div className="guide-section__title">
-            {editing ? (
-              <EuiFieldText
-                placeholder="title"
-                value={title}
-                onChange={(e) => updateSection("title", e.target.value, index)}
-              />
-            ) : (
-              <div>{title}</div>
-            )}
-            <EuiButtonIcon
-              aria-label="collapse-icon"
-              iconType={isCollapsed ? "arrowDown" : "arrowUp"}
-              iconSize="l"
-              size="m"
-              className="guide-section__title--collapse"
-              onClick={() => {
-                handleCollapse(index);
-              }}
-            ></EuiButtonIcon>
-          </div>
-          {!isCollapsed && (
-            <div className={`guide-section__body${editing ? " editing" : ""}`}>
-              {editing ? (
-                <EuiMarkdownEditor
-                  aria-label="Body markdown editor"
-                  value={body}
-                  onChange={(value) => updateSection("body", value, index)}
-                  height={400}
-                />
-              ) : (
-                <EuiMarkdownFormat>{body}</EuiMarkdownFormat>
-              )}
-            </div>
-          )}
-        </EuiPanel>
-      );
-
       return editing ? (
         <EuiDraggable
           spacing="m"
@@ -199,10 +97,26 @@ export const GuideDetailView: FunctionComponent<GuideDetailViewProps> = (): Reac
           draggableId={index.toString()}
           isDragDisabled={!editing}
         >
-          {sectionPanel}
+          <GuideDetailSection
+            title={title}
+            body={body}
+            index={index}
+            editing={editing}
+            isCollapsed={isCollapsed}
+            handleCollapse={handleCollapse}
+            updateSection={updateSection}
+          />
         </EuiDraggable>
       ) : (
-        sectionPanel
+        <GuideDetailSection
+          title={title}
+          body={body}
+          index={index}
+          editing={editing}
+          isCollapsed={isCollapsed}
+          handleCollapse={handleCollapse}
+          updateSection={updateSection}
+        />
       );
     });
   };
@@ -215,75 +129,21 @@ export const GuideDetailView: FunctionComponent<GuideDetailViewProps> = (): Reac
             className="guide-detail__controls"
             style={editing ? { paddingRight: 8 } : {}}
           >
-            {editing ? (
-              <>
-                <EuiButtonIcon
-                  aria-label="cancel"
-                  className="guide-controls__button"
-                  display="fill"
-                  iconType="trash"
-                  color="danger"
-                  onClick={handleCancel}
-                  size="m"
-                  iconSize="l"
-                />
-                <EuiButtonIcon
-                  className="guide-controls__button"
-                  display="fill"
-                  iconType="save"
-                  color="success"
-                  onClick={handleSave}
-                  size="m"
-                  iconSize="l"
-                />
-                <EuiButtonIcon
-                  className="guide-controls__button"
-                  display="fill"
-                  iconType="plus"
-                  color="success"
-                  onClick={addSection}
-                  size="m"
-                  iconSize="l"
-                />
-              </>
-            ) : (
-              <EuiButtonIcon
-                className="guide-controls__button"
-                display="fill"
-                iconType="pencil"
-                size="m"
-                iconSize="l"
-                onClick={() => setEditing(!editing)}
-              />
-            )}
+            <GuideDetailControls
+              editing={editing}
+              handleCancel={handleCancel}
+              handleSave={handleSave}
+              handleAddSection={handleAddSection}
+              handleSetEditing={handleSetEditing}
+            />
           </div>
+
           <div className="guide-detail__content">
-            <EuiPanel
-              className="guide-content__side-nav"
-              style={editing ? { marginTop: 8 } : {}}
-              hasShadow={false}
-              hasBorder
-            >
-              <EuiAvatar
-                size="xl"
-                name={mockGuide.title}
-                imageUrl={
-                  "https://www.textures-resource.com/resources/sheet_icons/4/3300.png"
-                }
-              ></EuiAvatar>
-              <EuiSideNav
-                className="guide-content__side-nav__items"
-                aria-label="Basic example"
-                style={{ width: 192 }}
-                items={[
-                  {
-                    name: mockGuide.title,
-                    id: 0,
-                    items: buildSideNaveItems(),
-                  },
-                ]}
-              />
-            </EuiPanel>
+            <GuideDetailSideNav
+              editing={editing}
+              title={mockGuide.title}
+              sections={sections}
+            />
             <div id="sections" className="guide-content__sections">
               {editing ? (
                 <EuiDragDropContext onDragEnd={handleDragEnd}>
